move routing to seperate file and folder
